pyproject.toml
[tool.poetry]
name = "docs-flow-backend"
version = "0.0.1"
description = ""
authors = ["Merser4"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
